machine_learning/requirements.production.txt
# Production requirements with pinned versions
# Last updated: December 6, 2025

# Core ML/Data
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.6.1
xgboost==2.0.3
lightgbm==4.1.0
imbalanced-learn==0.11.0

# Model Interpretation
shap==0.44.0

# API Framework
fastapi==0.109.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
python-multipart==0.0.6

# MLOps
mlflow==2.10.0
sqlalchemy==2.0.25

# Monitoring & Metrics
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# Rate Limiting
slowapi==0.1.9

# Error Tracking
sentry-sdk==1.40.0

# Database
psycopg2-binary==2.9.9

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==23.12.1
flake8==7.0.0
bandit==1.7.6
safety==2.3.5

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.4

# Utilities
python-dateutil==2.8.2
pytz==2023.3
joblib==1.3.2
